6_AI_MCP/wykład/requirements.txt
langchain-core
langchain-community
langchain
pandas
openai
python-dotenv
mcp
langgraph
langchain-mcp-adapters
langchain-openai